pyproject.toml
[tool.poetry]
name = "airtable-api"
version = "0.1.0"
description = ""
authors = ["mohamad-ali-nasser <mhd.ali.nasser@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
pandas = ">=2.3.1,<3.0.0"
pyairtable = ">=3.1.1,<4.0.0"
fastapi = ">=0.116.1,<0.117.0"
python-dotenv = ">=1.1.1,<2.0.0"
uvicorn = "^0.35.0"
pytest = "^8.4.1"
httpx = "^0.28.1"
openai = "^1.99.9"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
